Add authentication to the backend and controller services

+++ backend/uploads/views.py
@@ -1,9 +1,12 @@
 from django import views
 from django.http.response import JsonResponse
+from django.shortcuts import get_object_or_404
 from rest_framework.decorators import action
 from uploads.models import ModelUpload, Model
 from rest_framework import viewsets, serializers
 from rest_framework.response import Response
+from rest_framework.status import HTTP_403_FORBIDDEN, HTTP_400_BAD_REQUEST
+from dj_rest_auth.jwt_auth import JWTCookieAuthentication
 from uploads.errors import ErrorResponse
 from uploads.google_signing_helpers import (
     generate_upload_signed_url_v4,
@@ -16,7 +19,7 @@ BUCKET_NAME = "easytensor-model-uploads"
 class ModelSerializer(serializers.ModelSerializer):
     class Meta:
         model = Model
-        fields = ["name", "address", "size", "scale", "id", "deployed"]
+        fields = ["owner", "name", "address", "size", "scale", "id", "deployed"]
 
 
 class CommentSerializer(serializers.Serializer):
@@ -27,8 +30,8 @@ class CommentSerializer(serializers.Serializer):
 
 class ModelUploadViewSet(viewsets.ModelViewSet):
     queryset = ModelUpload.objects.all()
-    authentication_classes = []
-    permission_classes = []
+    authentication_classes = [JWTCookieAuthentication]
+    # permission_classes = []
 
     def create(self, request):
         print("hello?")
@@ -51,13 +54,16 @@ class ModelUploadViewSet(viewsets.ModelViewSet):
 class ModelViewSet(viewsets.ModelViewSet):
     queryset = Model.objects.all()
     serializer_class = ModelSerializer
-    authentication_classes = []
-    permission_classes = []
+    authentication_classes = [JWTCookieAuthentication]
 
     def create(self, request):
         from pprint import pprint
 
         pprint(request.data)
+        
+        data = request.data
+        if "owner" not in data:
+            data["owner"] = request.user.id
         serializer = ModelSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -67,8 +73,35 @@ class ModelViewSet(viewsets.ModelViewSet):
 
     @action(methods=["delete"], detail=False)
     def delete(self, request, *args, **kwargs):
+        if not request.user.is_staff:
+            return Response(status=HTTP_403_FORBIDDEN)
         Model.objects.all().delete()
-        return Response({})
+        return Response()
+
+    def list(self, request, *args, **kwargs):
+        print("request user:")
+        print(request.user)
+        print(request.COOKIES)
+        if request.user.is_staff:
+            models = Model.objects.all()
+        else:    
+            models = Model.objects.filter(owner=request.user.id)
+        serializer = ModelSerializer(models, many=True)
+        return Response(serializer.data)
+
+    def update(self, request, pk=None):
+        _ = get_object_or_404(Model, pk=pk, owner=request.user.id)
+        serializer = ModelSerializer(request.data)
+        if serializers.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
+
+    def retrieve(self, request, pk, *args, **kwargs):
+        model = get_object_or_404(Model, pk=pk, owner=request.user.id)
+        
+        serializer = ModelSerializer(model)
+        return Response(serializer.data)
 
 
 def health_check(request):
